Fix ESP-IDF v5.4 build for ESP32-P4 target
- Rename components/hal → components/board_hal to avoid shadowing
  IDF's built-in hal component (caused missing hal/sha_types.h etc.)
- Add main/idf_component.yml for managed components:
  esp_lcd_touch, esp_lcd_touch_gt911, lvgl/lvgl, mdns
- Update CMakeLists.txt in core, ui, main to reference board_hal
- board_hal/Display.cpp: remove non-existent relax_on_idle flag
- board_hal/Touch.cpp: replace C99 out-of-order designated initialiser
  with explicit member assignment (C++ compatible)
- core/Storage.cpp: replace sys/statvfs.h + statvfs() with
  esp_spiffs_info() (statvfs not available on ESP-IDF)
- ui/Widgets.cpp: lv_event_send → lv_obj_send_event (LVGL v9 API)
- core/BottomNav.h: move NavItem struct to public section so
  anonymous-namespace callback can access it
- services/TimeService: add hour() accessor (used by LauncherApp)
- services/OTAService.cpp: wrap esp_http_client_config_t in
  esp_https_ota_config_t (API changed in IDF v5.x)
- services/MarkdownParser.cpp: std::strncasecmp → ::strncasecmp
- apps/*: cast uint32_t/int32_t args to (unsigned)/(int) for %u/%d
  format specifiers (RISC-V: int32_t = long int, not int)
- apps/RecipesApp.cpp: dynamic_cast → static_cast (no RTTI needed)
- sdkconfig.defaults: add 16 MB flash size, LVGL Montserrat fonts
  (16/20/24/32/48pt), host WiFi (ESP32-C6 co-processor),
  disable hardware SHA (include-path issue with IDF v5.4/P4)

+++ components/apps/DishwasherApp.cpp
@@ -176,7 +176,9 @@ void DishwasherApp::refresh() {
     lv_label_set_text(status_label_, state_text(state_));
     lv_obj_set_style_text_color(status_label_, lv_color_hex(state_color(state_)), 0);
     char time_buf[16];
-    std::snprintf(time_buf, sizeof(time_buf), "%02u:%02u", remaining_sec_ / 60, remaining_sec_ % 60);
+    std::snprintf(time_buf, sizeof(time_buf), "%02u:%02u",
+                  static_cast<unsigned>(remaining_sec_ / 60),
+                  static_cast<unsigned>(remaining_sec_ % 60));
     lv_label_set_text(time_label_, time_buf);
     lv_arc_set_value(arc_, total_sec_ == 0 ? 0 : static_cast<int>((total_sec_ - remaining_sec_) * 100 / total_sec_));
 
@@ -207,7 +209,9 @@ void DishwasherApp::on_update(float delta_sec) {
             remaining_sec_ = remaining_sec_ > 1 ? remaining_sec_ - 1 : 0;
             if (time_label_) {
                 char buf[16];
-                std::snprintf(buf, sizeof(buf), "%02u:%02u", remaining_sec_ / 60, remaining_sec_ % 60);
+                std::snprintf(buf, sizeof(buf), "%02u:%02u",
+                              static_cast<unsigned>(remaining_sec_ / 60),
+                              static_cast<unsigned>(remaining_sec_ % 60));
                 lv_label_set_text(time_label_, buf);
             }
             if (arc_ && total_sec_ > 0)
@@ -222,4 +226,3 @@ void DishwasherApp::on_update(float delta_sec) {
 }
 
 } // namespace apps
-
